# code/backend/docs/openapi.yaml
openapi: 3.0.3
info:
  title: Gestor Job API
  version: 0.6.0
  description: |
    API do Gestor Job (SaaS multi-tenant para agências).
    F6 — Health com check de database (503 se falhar); deploy CI/self-hosted.
    F5 — Policies por recurso (403 papel / 404 não visível), headers CSP, CORS só FRONTEND_URL em prod, throttle login.
    F3 — Kanban + timer por fase (Análise→Produção→Revisão→Correção). Atraso = badge.
servers:
  - url: http://localhost:8000/api/v1
...
components:
  securitySchemes:
    sanctum:
      type: http
      scheme: bearer
...
paths:
...
  /empresas/{empresa}/convite:
    post:
      operationId: reenviarConvite
      tags: [Empresas]
      summary: Reenvia convite se a conta ainda não foi ativada
      security: [{ sanctum: [] }]
      parameters:
        - in: path
          name: empresa
          required: true
          schema: { type: integer }
      responses:
        '200': { description: Novo convite_url }
        '403': { description: Não é super_admin }
        '409': { description: Conta já ativada }
  /convites/{token}:
    get:
      operationId: previewConvite
      tags: [Auth]
      summary: Dados públicos do convite (sem senha)
      parameters:
        - in: path
          name: token
          required: true
          schema: { type: string }
      responses:
        '200': { description: Nome, e-mail, empresa }
        '404': { description: Token inválido }
        '422': { description: Expirado }
    post:
      operationId: aceitarConvite
      tags: [Auth]
      summary: Admin define senha e ativa a conta
      parameters:
        - in: path
          name: token
          required: true
          schema: { type: string }
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [name, password, password_confirmation]
              properties:
                name: { type: string }
                password: { type: string, minLength: 8 }
                password_confirmation: { type: string }
      responses:
        '200': { description: Token Sanctum + user }
        '404': { description: Token inválido }
        '422': { description: Expirado ou senha inválida }
...